move createOutputsChild and its helper functions into presenter js for a better seperation
